requirements.txt
ccxt==4.1.13
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
ta==0.10.2
requests==2.31.0
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
pymongo==4.6.1
motor==3.3.2 